new_installer: fix the issue reported by ajgenius
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/apps/new_installer@12942 67a373e5-eb02-0410-a15c-ee090a768436

+++ res/unix/bin/install_plugins.sh
@@ -4,8 +4,11 @@ INSTALL_PATH="${INSTALL_PATH:-$PWD}"
 
 cd "$INSTALL_PATH"
 
-# Register plugins
-echo "pluginmanager.loadplugin=`cat plug`" >> freenet.ini
+if test -e plug
+then
+	# Register plugins
+	echo "pluginmanager.loadplugin=`cat plug`" >> freenet.ini
 
-# Cleanup
-rm -f plug plug2
+	# Cleanup
+	rm -f plug plug2
+fi
